Add files via upload
This repository hosts a Power BI project that is dedicated to developing a sales dashboard for a company. The main objective of this project is to provide an in-depth analysis of sales data, enabling users to gain valuable insights into the company's sales performance. By utilizing the powerful features of Power BI, the dashboard aims to present the data in a visually appealing and user-friendly manner, making it easier for users to interpret and understand the information.